+++ clara/transpiler/naive_bayes/__init__.py
@@ -1,0 +1,1 @@
+from.gausianNB import GaussianNBTranspiler